requirements.txt
tensorflow==2.16.1
numpy
matplotlib
opencv-python
pillow